pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = ""
authors = ["scmopt <kubomikio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
django-crispy-forms = "^1.11.2"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
